chore(vector stores, transports): move inline test modules to sibling files
